backend/pyproject.toml
[project]
name = "openar-backend"
version = "0.1.0"
description = "Backend for boat detection using Roboflow Inference"
requires-python = ">=3.9"
dependencies = [
    "inference-cli",
    "inference",
    "inference-sdk",
    "roboflow",
    "supervision",
    "opencv-python-headless",
    "pillow",
    "numpy",
    "scipy",
    "ultralytics",
    "lap>=0.5.12",
    "boto3",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.30",
    "psycopg[binary]>=3.2.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "redis>=5.0.8",
    "torch",
    "torchvision",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu121", marker = "sys_platform != 'darwin'" }]
torchvision = [{ index = "pytorch-cu121", marker = "sys_platform != 'darwin'" }]
